update open order success

+++ lib/services/firestoreDB/order_owner_db_service.dart
@@ -1,12 +1,15 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_application_1/src/features/auth/models/order_owner.dart';
+import 'package:flutter_application_1/src/routing/routes_const.dart';
 
 class OrderOwnerDatabaseService{
   final FirebaseFirestore _db = FirebaseFirestore.instance;
   final CollectionReference orderCollection = FirebaseFirestore.instance.collection('open order');
   //add order
-  addOrder(OrderOwnerModel orderData) async{
-    await _db.collection('open order').add(orderData.toOrderOwnerJason());
+  Future<DocumentReference>addOrder(OrderOwnerModel orderData) async{
+    DocumentReference documentReference = await _db.collection('open order').add(orderData.toOrderOwnerJason());
+    return documentReference;
   }
 
   //update order
@@ -15,8 +18,39 @@ class OrderOwnerDatabaseService{
   }
 
   //delete order
-  Future<void> deleteorder(String documentId) async{
-    await _db.collection('open order').doc(documentId).delete();
+  Future<void> deleteOrder(String? documentId, BuildContext context) async{
+    if (documentId == null || documentId.isEmpty) {
+      // Show an alert or return an appropriate response
+      showDialog(
+        context: context,
+        builder: (context) {
+          return AlertDialog(
+            title: const Text('Invalid Document ID'),
+            content: const Text('The document ID is null or empty. Cannot delete.'),
+            actions: [
+              TextButton(
+                onPressed: () {
+                  Navigator.of(context).pop(); // Close the dialog
+                },
+                child: const Text(
+                  'OK',
+                  style: TextStyle(
+                  fontSize: 20
+                )
+                ),
+              ),
+            ],
+          );
+        },
+      );
+    } else {
+      // Call the deletePayment function with a valid documentId
+      await _db.collection('open order').doc(documentId).delete();
+      Navigator.of(context).pushNamedAndRemoveUntil(
+        orderAddPageRoute, 
+        (route) => false,
+      );
+    } 
   }
 
   Stream<List<OrderOwnerModel>> getOrderMethods(){
@@ -34,18 +68,4 @@ class OrderOwnerDatabaseService{
     );
   }
 
-  //get specific order
-  // Future<OrderOwnerModel> getOrder(String documentId) async{
-  //   try{
-  //     DocumentSnapshot<Map<String, dynamic>> snapshot = await _db.collection('open order').doc(documentId).get();
-  //     if(snapshot.exists){
-  //       return OrderOwnerModel.fromDocumentSnapshot(snapshot);
-  //     }else{
-  //       return OrderOwnerModel.defaults();
-  //     }
-  //   }catch(e){
-  //     print('Error fetching order');
-  //     throw Exception('Error fetching menu');
-  //   }
-  // }
 }
